Split header and footer to own components and create data
